HireMateChatbot/requirements.txt
openai
gspread
google-auth
google-api-python-client
langchain